+++ MordoEngine/src/Renderer/Renderer.h
@@ -1,8 +1,9 @@
 #pragma once
+#include "../Core/Shader/Shader.h"
 
 class Renderer
 {
 public:
-	virtual void Render() = 0;
+	virtual void Render(const Shader& shader) = 0;
 	virtual ~Renderer() = default;
 };